Move to network.js as tokens.js is now redundant
