+++ sync/network.js
@@ -1,3 +1,5 @@
+const crypto = require('crypto');
+
 let network = {}
 
 function addNetworkNode(id, token) {
@@ -14,6 +16,10 @@ function removeNetworkNode(id) {
     }
 }
 
+function hasNetworkNode(id) {
+    return network[id] != undefined
+}
+
 function getNetworkNodes() {
     return network
 }
@@ -22,10 +28,29 @@ function setNetworkNodes(recvNodes) {
     network = recvNodes
 }
 
+function getOwnToken() {
+    // console.log(network[process.env.DATASYNC_OWN_ID].token)
+    return network[process.env.DATASYNC_OWN_ID].token
+}
+function getInstanceToken(id) {
+    if (network[id] != undefined) {
+        return network[id].token
+    } else {
+        return ''
+    }
+}
+function generateToken() {
+    return crypto.randomBytes(process.env.DATASYNC_TOKEN_BYTES || 256).toString('base64');
+}
+
 module.exports = {
     addNetworkNode,
     getNetworkNode,
     removeNetworkNode,
+    hasNetworkNode,
     getNetworkNodes,
-    setNetworkNodes
+    setNetworkNodes,
+    getOwnToken,
+    getInstanceToken,
+    generateToken
 }
